refactor: 매핑 클래스 디렉토리 구조 변경 #91
