+++ src/main/java/SMU/BAMBOO/Hompage/domain/tag/entity/Tag.java
@@ -1,6 +1,6 @@
 package SMU.BAMBOO.Hompage.domain.tag.entity;
 
-import SMU.BAMBOO.Hompage.domain.mapping.LibraryPostTag;
+import SMU.BAMBOO.Hompage.domain.mapping.libraryPostTag.LibraryPostTag;
 import SMU.BAMBOO.Hompage.domain.tag.dto.TagRequestDTO;
 import SMU.BAMBOO.Hompage.global.common.BaseEntity;
 import jakarta.persistence.*;
