+++ src/test/java/SMU/BAMBOO/Hompage/domain/libraryPost/entity/LibraryPostTest.java
@@ -1,7 +1,7 @@
 package SMU.BAMBOO.Hompage.domain.libraryPost.entity;
 
 import SMU.BAMBOO.Hompage.domain.libraryPost.dto.LibraryPostRequestDTO;
-import SMU.BAMBOO.Hompage.domain.mapping.LibraryPostTag;
+import SMU.BAMBOO.Hompage.domain.mapping.libraryPostTag.LibraryPostTag;
 import SMU.BAMBOO.Hompage.domain.tag.entity.Tag;
 import org.junit.jupiter.api.Test;
